+++ packages/subgraph/scripts/render-manifest.ts
@@ -38,19 +38,27 @@ const envNameToConfig: Record<
     },
     "testing-5-0": {
       network: "goerli",
-      startBlock: 9618015 // goerli, block number when protocol was deployed https://goerli.etherscan.io/address/0xDA01E28EA66B4294EeE21013bC45a0114E976Da4
+      startBlock: 9757131 // goerli, block number when protocol was deployed https://goerli.etherscan.io/tx/0x3c51e116a6a1849936e6bd45fcee78017259ce86a93601e5e7c77cbdc25b6f38
     }
   },
   staging: {
     "staging-80001-0": {
       network: "mumbai",
       startBlock: 28566747 // mumbai, block number when protocol was deployed
+    },
+    "staging-5-0": {
+      network: "goerli",
+      startBlock: 9756834 // goerli, block number when protocol was deployed https://goerli.etherscan.io/tx/0x7d4b731fc8b9fe77999f46a52891f2bd13651666bb9a15d7044b33bddd52b355
     }
   },
   production: {
     "production-137-0": {
       network: providers.getNetwork(chainId).name,
       startBlock: 34258150 // polygon, block num when protocol is deployed
+    },
+    "production-1-0": {
+      network: providers.getNetwork(chainId).name,
+      startBlock: 99999999999999 //TODO:  ethereum, block num when protocol is deployed
     }
   }
 };